nse/requirements.txt
fastapi==0.63.0
uvicorn==0.15.0
tortoise-orm==0.16.13
python-multipart==0.0.5
PyJWT==1.7.1
aiohttp==3.9.3
aiodns==2.0.0
python-dotenv==0.15.0
flake8==3.9.2
ipdb==0.13.9
ipython==7.17.0
brotlipy==0.7.0
cryptography
aiomysql==0.2.0
pandas==2.0.3
requests==2.31.0
fyers-apiv3
pydantic[email]